fixed warnings

+++ tutorials/viewer_anim/viewer_anim_device.cpp
@@ -396,7 +396,7 @@ inline Vec3fa face_forward(const Vec3fa& dir, const Vec3fa& _Ng) {
       for (unsigned int i=0; i<g_ispc_scene->numLights; i++)
       {
         /* init shadow/occlusion rays */
-        for (size_t n=0;n<N;n++)
+        for (int n=0;n<N;n++)
         {
           RTCRay& ray = rays[n];
           const bool valid = ray.geomID != RTC_INVALID_GEOMETRY_ID;
